feat: добавлен BaseOperationController
